Module solana_send_transaction_service::send_transaction_service

Structs

Config

CurrentLeaderInfo

A struct responsible for holding up-to-date leader information used for sending transactions.

ProcessTransactionsResult 🔒

SendTransactionService

SendTransactionServiceStats 🔒

Metrics of the send-transaction-service.

SendTransactionServiceStatsReport 🔒

TransactionInfo

Constants

DEFAULT_BATCH_SEND_RATE_MS 🔒

Default maximum batch waiting time in ms. If this time is reached, whatever transactions are cached will be sent.

DEFAULT_LEADER_FORWARD_COUNT 🔒

Default number of leaders to forward transactions to

DEFAULT_RETRY_RATE_MS 🔒

Default retry interval

DEFAULT_SERVICE_MAX_RETRIES 🔒

Default max number of time the service will retry broadcast

DEFAULT_TRANSACTION_BATCH_SIZE 🔒

Default batch size for sending transaction in batch When this size is reached, send out the transactions.

LEADER_INFO_REFRESH_RATE_MS

The leader info refresh rate.

MAX_BATCH_SEND_RATE_MS

MAX_RETRY_SLEEP_MS

The maximum duration the retry thread may be configured to sleep before processing the transactions that need to be retried.

MAX_TRANSACTION_BATCH_SIZE

MAX_TRANSACTION_QUEUE_SIZE 🔒

Maximum size of the transaction queue

MAX_TRANSACTION_SENDS_PER_SECOND

Maximum transaction sends per second

SEND_TRANSACTION_METRICS_REPORT_RATE_MS 🔒

Report the send transaction memtrics for every 5 seconds.